backend/pyproject.toml
# Using a proper build system makes uv install the package itself
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[project]
name = "crypkit-tracker"
version = "0.1.0"
authors = [
  { name = "CrypKit Tracker" },
]
description = "Cryptocurrency portfolio tracker using FastAPI, PostgreSQL, and Redis"
requires-python = ">=3.12"
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi>=0.101.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0.39",
    "pydantic>=2.10.6",
    "redis>=5.2.1",
    "psycopg2-binary>=2.9.10",
    "asyncpg>=0.30.0",
    "aiohttp>=3.11.14",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.0",
    "alembic>=1.15.1",
    "pydantic-settings>=2.8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.0.0",
    "aiosqlite>=0.21.0",  # For SQLite-based testing
    "ruff>=0.11.2"
]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 120
indent-width = 4

target-version = "py312"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"
